+++ src/rpc_client/mod.rs
@@ -5,6 +5,7 @@ pub mod beacon_ops;
 pub mod chain_ops;
 pub mod common_ops;
 pub mod eth_ops;
+pub mod gas_ops;
 pub mod net_ops;
 pub mod node_ops;
 pub mod state_ops;